\name{ordijuice-package}
\alias{ordijuice-package}
\alias{ordijuice}
\docType{package}
\title{
Unconstrained ordinations of vegetation data in JUICE program
}
\description{
R library not used for end user, but only to be run from
        JUICE program
}
\details{
\tabular{ll}{
Package: \tab ordijuice\cr
Type: \tab Package\cr
Version: \tab 3.3.8\cr
Date: \tab 2010-06-21\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
David Zeleny

Maintainer: David Zeleny <zeleny@sci.muni.cz>
}
\references{
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
